Add files via upload
This first repo is probably going to be a  bunch of small random things tbh. I have a bunch of small python scripts from where I was practicing and studying things but most isnt good enough to show here. What I show here are things I'm actually happy with and proud to show off, even if it's really small and basic.